Agentic AI Judge/requirements.txt
transformers
accelerate
huggingface_hub
pyyaml
tqdm
scipy
pandas
torch
